Commited to branch to save

+++ src/tcp/tcp_op.h
@@ -23,7 +23,8 @@ extern uint32_t PKT_SIZE;
 extern char *VARIANT;
 extern uint64_t ERTT;
 extern uint64_t TIMEOUT;
-extern uint32_t TESTING_PERIOD;
+extern uint64_t TESTING_PERIOD;
+
 struct tcp_packet_entry
 {
   TAILQ_ENTRY (tcp_packet_entry) entry;
